Changed folders

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    double calculate(double meters, int kilograms) {
+    int calculate(double meters, int kilograms) {
         double result;
         result = kilograms / (meters * meters);
         return (int) result;
